+++ lib/src/gcc.dart
@@ -1,7 +1,7 @@
 part of ccompilers.ccompilers;
 
 class Gcc extends CommandLineTool {
-  Gcc() {
+  Gcc({Logger logger}) : super(logger: logger) {
     executable = 'gcc';
   }
 }