+++ lib/src/mslink.dart
@@ -1,7 +1,7 @@
 part of ccompilers.ccompilers;
 
 class Mslink extends MsvsTool {
-  Mslink({int bits}) : super(bits: bits) {
+  Mslink({int bits, Logger logger}) : super(bits: bits, logger: logger) {
     executable = 'link.exe';
   }
 }