+++ lib/src/msvc.dart
@@ -1,7 +1,7 @@
 part of ccompilers.ccompilers;
 
 class Msvc extends MsvsTool {
-  Msvc({int bits}) : super(bits: bits) {
+  Msvc({int bits, Logger logger}) : super(bits: bits, logger: logger) {
     executable = 'cl.exe';
   }
 }