+++ lib/src/msvs_tool.dart
@@ -3,7 +3,7 @@ part of ccompilers.ccompilers;
 abstract class MsvsTool extends CommandLineTool {
   int _bits;
 
-  MsvsTool({int bits}) {
+  MsvsTool({int bits, Logger logger}) : super(logger: logger) {
     if (!(bits == null || bits == 32 || bits == 64)) {
       throw new ArgumentError('bits: $bits');
     }